requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cpf-cnpj==1.0.0
django-crispy-forms==2.0
django-discover-runner==1.0
django-role-permissions==3.1.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
